% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/injector.R
\name{binder}
\alias{binder}
\title{Binder factory}
\usage{
binder(parent = .binder, callback = function(binder) binder)
}
\arguments{
\item{parent}{of the new binder, injection will propagate up the 
parent stack looking for keys; if omitted defaults to root binder}

\item{callback}{called with the newly created binder and the
result is returned; if omitted just the new binder is returned}
}
\value{
result of the injected callback if one is specified,
otherwise the new binder
}
\description{
Binder factory
}
\examples{
b <- binder ()
}

