% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/injector.R
\name{default}
\alias{default}
\title{Default scope, bindings are provisioned each time a bean is
injected}
\usage{
default(key, provider)
}
\arguments{
\item{key}{of the new binding}

\item{provider}{unscoped delegate, no argument function responsible
for provision}
}
\description{
Default scope, bindings are provisioned each time a bean is
injected
}

