% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/injectoR.R
\name{inject}
\alias{inject}
\title{Injects the callback function}
\usage{
inject(callback, binder = .binder)
}
\arguments{
\item{callback}{function to inject, a function accepting arguments
to be matched to injectable keys}

\item{binder}{containing the injectables}
}
\value{
result of the injected callback evaluation
}
\description{
Injects the callback function
}

