% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intg1.r
\name{intg1}
\alias{intg1}
\title{Integration - 1}
\usage{
intg1(reference, query)
}
\arguments{
\item{reference}{A Seurat object obtained from load_ref function}

\item{query}{A Seurat object obtained from load_query function}
}
\value{
A Seurat object, where the reference and query are integrated
}
\description{
A function that integrate both Seurat objects reference and query.
The function use the 'reference' obtained from load_ref() and
the 'query' obtained from load_query(). This function includes
preprocessing, normlization, search of variables features and anchors,
for data integration of Seurat package workflow.
}
