% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_ref.r
\name{load_ref}
\alias{load_ref}
\title{Loading reference}
\usage{
load_ref(ref)
}
\arguments{
\item{ref}{Path to the input file (.h5seurat)}
}
\value{
A Seurat object, with trajectory coordinates as metadata
}
\description{
This function loads a reference Seurat object with trajectory
inference coordinates as metadata. It assumes that the 'Seurat'
object was previously generate by using CreateSeuratObject (Seurat
package). Trajectory coordinates (comp_1 and comp_2) were obtained
with one dynverse package methods and then added to the 'Seurat' object,
by using AddMetaData (Seurat package). If the reference
have < 20000 cells, it will use all of them. If it is between 20000
and 25000, it will randomly select half of the data. And if it have > 25000
cells, it will randomly select one third of the data.
}
