% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/injectoR.R
\name{proxy}
\alias{proxy}
\title{Proxy scope applicable for function type beans only, provisioned
once for each use allowing for circular dependencies}
\usage{
proxy(key, provider)
}
\arguments{
\item{key}{of the binding}

\item{provider}{unscoped delegate, no argument function responsible
for provision}
}
\description{
Proxy scope applicable for function type beans only, provisioned
once for each use allowing for circular dependencies
}

