% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transf_coord.r
\name{transf_coord}
\alias{transf_coord}
\title{Transfering coordinates}
\usage{
transf_coord(reference, dims, n_comps, integrated1, data_name)
}
\arguments{
\item{reference}{A Seurat object obtained from load_ref function}

\item{dims}{number of PCA dimensions used to calculate the reference's UMAP}

\item{n_comps}{number of components of your trajectory (2 or 3)}

\item{integrated1}{A Seurat object obtained from intg1 function}

\item{data_name}{name of the data (same as used in load_query)}
}
\value{
A Seurat object, with only query data and transfered
trajectory coordinates
}
\description{
A function that transfer the Trajectory coordinates
(comp_1, and comp_2, or and comp_3) within the metadata
of 'reference' Seurat object. The function use Seurat package
workflow of mapping and annotating query datasets.
}
\details{
the 'reference' obtained from load_ref() and
the 'query' obtained from load_query(). This function includes
preprocessing, normlization, search of variables features and anchors,
for data integration of Seurat package workflow.
}
