% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trj_inf.r
\name{trj_inf}
\alias{trj_inf}
\title{Trajectory injection}
\usage{
trj_inf(integrated2, n_comps, model_rooted, root)
}
\arguments{
\item{integrated2}{A Seurat object obtained from transf_coord function}

\item{n_comps}{number of components of your trajectory (2 or 3)}

\item{model_rooted}{Path to the trajectory model (.Robj)}

\item{root}{name of the milestone used for rooting the reference trajetory}
}
\description{
trajectory model obtain by infer_trajectory (dynwrap package)
}
