% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/injectoR.R
\name{default}
\alias{default}
\title{Default scope, bindings are provisioned each time a bean is
injected}
\usage{
default(key, provider)
}
\description{
Default scope, bindings are provisioned each time a bean is
injected
}

