% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/injector.R
\name{default}
\alias{default}
\title{Default scope, bindings are provisioned each time a bean is
injected}
\usage{
default(provider)
}
\arguments{
\item{provider}{unscoped delegate, no argument function responsible
for provision}
}
\description{
Default scope, bindings are provisioned each time a bean is
injected
}

