% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/injector.R
\name{inject}
\alias{inject}
\title{Injects the callback function}
\usage{
inject(callback, binder = .binder)
}
\arguments{
\item{callback}{function to inject, a function accepting arguments
to be matched to injectable keys; no errors are thrown if no binding
is found for a key, this is the intended mechanic for optional
injection, if the callback is able to deal with a missing argument
the argument becomes optional}

\item{binder}{containing the injectables, defaults to root binder if
omitted}
}
\value{
result of the injected callback evaluation
}
\description{
Injects the callback function
}

