% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/injectoR.R
\name{singleton}
\alias{singleton}
\title{Singleton scope, bindings of this scope are provided once, on
initial demand}
\usage{
singleton(key, provider)
}
\arguments{
\item{key}{of the new binding}

\item{provider}{unscoped delegate, no argument function responsible
for provision}
}
\description{
Singleton scope, bindings of this scope are provided once, on
initial demand
}

