% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/injector.R
\name{singleton}
\alias{singleton}
\title{Singleton scope, bindings of this scope are provided once, on
initial demand}
\usage{
singleton(provider)
}
\arguments{
\item{provider}{unscoped delegate, no argument function responsible
for provision}
}
\description{
Singleton scope, bindings of this scope are provided once, on
initial demand
}
\examples{
define (three = function () 3, scope = singleton, binder = binder ())
}

